split default config into system and web3
